Added two scripts for converting text to handwriting using the command line.
One script converts a hard coded line of text to handwriting. Another takes a text file as input and converts all the lines to handwriting, each line being separated by a new-line character. Selenium web automation is used to achieve this. Furthermore, the chromebrowser used has been made hidden, i.e., the script runs the tasks in the background, and hence gives the illusion of a feel of using the command line to get the desired result. Further improvements to the script are welcome, as only a few of the options have been scripted.